Fix build directory cleaner, to not follow links on Windows.
BUG=
R=ricow@google.com

Review URL: https://codereview.chromium.org//1219833003.

+++ tools/clean_output_directory.py
@@ -12,7 +12,13 @@ import utils
 def Main():
   build_root = utils.GetBuildRoot(utils.GuessOS())
   print 'Deleting %s' % build_root
-  shutil.rmtree(build_root, ignore_errors=True)
+  if sys.platform != 'win32':
+    shutil.rmtree(build_root, ignore_errors=True)
+  else:
+    # Intentionally ignore return value since a directory might be in use.
+    subprocess.call(['rmdir', '/Q', '/S', build_root],
+                    env=os.environ.copy(),
+                    shell=True)
   return 0
 
 if __name__ == '__main__':
